scheduler and executor share the same events decorator pattern
